+++ src/pymetamodels/__init__.py
@@ -1,7 +1,6 @@
 #!/usr/bin/python
-# FLC 2019
 
 from pymetamodels.pymeta_models import metamodel, load, newconf
 
-__version__ = "0.0.2"
+__version__ = "0.0.3"