+++ src/pymetamodels/clsxplots/obj_func.py
@@ -1,0 +1,13 @@
+#!/usr/bin/python
+
+def test_empty(y):
+    # Test for empty variable as []
+    
+    if type(y) == type([]):
+        if len(y) == 0:
+            return False
+        else:
+            return True
+    else:
+        return True
+